+++ backend/canisters/community/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support muting of @everyone notifications ([#8496](https://github.com/open-chat-labs/open-chat/pull/8496))
 - Support muted=false && @everyone_muted=true ([#8502](https://github.com/open-chat-labs/open-chat/pull/8502))
+- Remove legacy group membership fields from backend ([#8508](https://github.com/open-chat-labs/open-chat/pull/8508))
 
 ### Fixed